+++ packages/react/src/themes/teams-high-contrast/components/Chat/chatMessageVariables.ts
@@ -5,7 +5,7 @@ export default (siteVars: any): Partial<ChatMessageVariables> => {
     backgroundColor: siteVars.black,
     backgroundColorMine: siteVars.black,
     color: siteVars.white,
-    contentFocusOutlineColor: siteVars.yellow,
+    contentFocusOutlineColor: siteVars.colors.yellow[900], // Red flag (should this be accessibleYellow?)
     border: `1px solid ${siteVars.white}`,
   }
 }