+++ packages/react/src/themes/teams/components/RadioGroup/radioGroupItemVariables.ts
@@ -11,13 +11,13 @@ export type RadioGroupItemVariables = {
 }
 
 export default (siteVars: any): RadioGroupItemVariables => ({
-  color: siteVars.brand,
-  colorChecked: siteVars.white,
+  color: siteVars.colors.primary[500],
+  colorChecked: siteVars.colors.white,
   colorDisabled: siteVars.gray06,
 
-  colorBorder: siteVars.brand,
-  colorBorderChecked: siteVars.white,
+  colorBorder: siteVars.colors.primary[500],
+  colorBorderChecked: siteVars.colors.white,
 
-  colorBackground: siteVars.white,
-  colorBackgroundChecked: siteVars.brand,
+  colorBackground: siteVars.colors.white,
+  colorBackgroundChecked: siteVars.colors.primary[500],
 })