+++ packages/react/src/themes/teams/components/Tree/treeTitleVariables.ts
@@ -4,6 +4,6 @@ export interface TreeTitleVariables {
 
 export default (siteVars: any): TreeTitleVariables => {
   return {
-    defaultColor: siteVars.black,
+    defaultColor: siteVars.colors.grey[900],
   }
 }